fix: fix deployments scripts

+++ deploy_local.sh
@@ -6,7 +6,11 @@ make
 cd ../..
 hugo
 cd ../site-assets
-rm -rf ./*
-cp -r ../code/public/* .
-cp -r ../code/CNAME .
-cd ../code
+if [ $? -ne 0 ]; then
+	echo "You don't have the ../site-assets directory"
+else
+	rm -rf ./*
+	cp -r ../www/public/* .
+	cp -r ../www/CNAME .
+	cd ../www
+fi
